service-ecommerce/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
django-cors-headers==4.3.1
requests==2.31.0
python-dotenv==1.0.0
setuptools>=60.0.0
psycopg2-binary==2.9.7 